// src/app/models/eventoModel.ts
export interface Evento{
    id_evento: string;
    nm_Evento: string;
    desc_evento: string;
    dt_InicioEvento: string;
    dt_FimEvento: string;
    dt_Cadastro: string;
    id_RespEvento: string;
    id_Cidade: string;
    desc_endereco: string;
    vlr_ingresso: number;
    status_evento: string;
    idade_minima: number;
}